square boss

+++ game/logic/ship/components.h
@@ -80,7 +80,7 @@ struct Health : ecs::component {
   std::uint32_t max_hp = hp;
 
   std::uint32_t hit_timer = 0;
-  std::optional<std::uint32_t> hit_flash_ignore_index;
+  std::optional<std::size_t> hit_flash_ignore_index;
 
   std::optional<sound> hit_sound0 = sound::kEnemyHit;
   std::optional<sound> hit_sound1 = sound::kEnemyHit;
